Rename strip html transform to fit naming convention
