+++ transformer/transforms/string/strip_html.py
